requirements.txt
django==5.0.6
mysqlclient==2.2.4
pillow==10.3.0
psycopg==3.1.18
pytz==2024.1
sqlparse==0.5.0
django-filter==24.2
django-money==3.4.1
django-datetimepicker==3.14
xhtml2pdf==0.2.15
pdfkit==1.0.0
reportlab==4.0.4
py-moneyed==3.0
python-barcode==0.15.1